cleanup

+++ tests/functional/swiglu_test.py
@@ -49,7 +49,7 @@ def _generate_args(function: Callable) -> list:
     args += list(
         product(
             [(4100, 3700)],
-            [torch.float32, torch.bfloat16],
+            [torch.float32, torch.float16],
             [KernelBackend.mps],
             [function],
         )
